requirements.txt
r2pipe==0.9.9
web.py
pywin32
OTXv2
pefile
lief
networkx==2.2
yara-python
pydot
graphviz
spam_lists